requirements.txt
absl-py==0.8.1
altair==4.1.0
astor==0.8.0
astroid==2.3.1
astropy==3.2.2
atomicwrites==1.3.0
attrs==19.2.0
backcall==0.1.0
base58==2.0.1
bleach==3.1.0
blinker==1.4
boto3==1.14.12
botocore==1.17.12
cachetools==4.1.0
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
confuse==1.0.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
htmlmin==0.1.12
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.14.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numba==0.46.0
numpy==1.17.2
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.1
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
ptyprocess==0.6.0
py==1.8.0
pydeck==0.4.0b2
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.1
pytest-pylint==0.14.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
s3transfer==0.3.3
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
streamlit==0.62.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.1
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.2
twilio==6.32.0
typed-ast==1.4.0
tzlocal==2.1
urllib3==1.25.6
validators==0.15.0
watchdog==0.10.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
